fix: scheduling filters title style
JIRA: F1-437
risk: low
